+++ src/main/java/com/bb/reference/MobileLocationListItem.java
@@ -2,8 +2,8 @@ package com.bb.reference;
 
 public class MobileLocationListItem {
     private Long locationId;
-    private Float latitude;
-    private Float longitude;
+    private Double latitude;
+    private Double longitude;
     private String name;
     private String address;
 
@@ -15,19 +15,19 @@ public class MobileLocationListItem {
         this.locationId = locationId;
     }
 
-    public Float getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Float latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Float getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Float longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
